# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VK Bot API",
        "description": "API для управления VK ботом",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/cyclic-tasks": {
            "get": {
                "tags": [
                    "Cyclic Tasks"
                ],
                "summary": "Получить список циклических задач",
                "description": "Возвращает список всех циклических задач",
                "operationId": "08d002e5f84ce2950ea40dfc974f858e",
                "responses": {
                    "200": {
                        "description": "Список циклических задач успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "account_id": {
                                                        "type": "integer",
                                                        "example": 123
                                                    },
                                                    "total_task_count": {
                                                        "type": "integer",
                                                        "example": 100
                                                    },
                                                    "remaining_tasks_count": {
                                                        "type": "integer",
                                                        "example": 75
                                                    },
                                                    "tasks_per_hour": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "enum": [
                                                            "active",
                                                            "pause",
                                                            "done"
                                                        ],
                                                        "example": "active"
                                                    },
                                                    "selected_times": {
                                                        "type": "object",
                                                        "example": "{\"monday\": [true, false, true]}"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Список циклических задач получен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Произошла ошибка при получении списка задач"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/cyclic-tasks/delete-all-cyclic-tasks": {
            "delete": {
                "tags": [
                    "Cyclic Tasks"
                ],
                "summary": "Удалить все циклические задачи",
                "description": "Удаляет все циклические задачи из системы",
                "operationId": "de2e58e5a8df05a9e2306a4ba73c2cb1",
                "responses": {
                    "200": {
                        "description": "Все циклические задачи успешно удалены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Все циклические задачи удалены"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Произошла ошибка при удалении всех задач"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}